Moved ParamTable to OpTk
