+++ hat/backends/jextracted/shared/src/main/java/hat/backend/jextracted/JExtractedBackend.java
@@ -98,10 +98,10 @@ public abstract class JExtractedBackend extends JExtractedBackendDriver {
         if (transform) {
             System.out.println("COMPUTE entrypoint before injecting buffer tracking...");
             System.out.println(returnFOW.op.toText());
-            returnFOW = FuncOpWrapper.transformInvokes(prevFOW.lookup, prevFOW.op,(bldr, invokeOW) -> {
+            returnFOW = OpTk.transformInvokes(prevFOW.lookup, prevFOW.op,(bldr, invokeOW) -> {
                 CopyContext bldrCntxt = bldr.context();
                 //Map compute method's first param (computeContext) value to transformed model
-                Value cc = bldrCntxt.getValue(prevFOW.parameter(0));
+                Value cc = bldrCntxt.getValue(prevFOW.paramTable.list().getFirst().parameter);
                 if (invokeOW.isIfaceMutator()) {                    // iface.v(newV)
                     Value iface = bldrCntxt.getValue(invokeOW.op.operands().getFirst());
                     bldr.op(JavaOp.invoke(MUTATE.pre, cc, iface));  // cc->preMutate(iface);
