+++ hat/core/src/main/java/hat/optools/OpTk.java
@@ -25,6 +25,8 @@
 package hat.optools;
 
 import hat.callgraph.CallGraph;
+import hat.util.BiMap;
+import jdk.incubator.code.Block;
 import jdk.incubator.code.Op;
 import jdk.incubator.code.OpTransformer;
 import jdk.incubator.code.Quoted;
@@ -35,7 +37,9 @@ import jdk.incubator.code.dialect.java.ClassType;
 import jdk.incubator.code.dialect.java.JavaOp;
 import jdk.incubator.code.dialect.java.JavaType;
 import jdk.incubator.code.dialect.java.MethodRef;
+import jdk.incubator.code.dialect.java.PrimitiveType;
 
+import java.lang.foreign.GroupLayout;
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
@@ -44,10 +48,12 @@ import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.BiFunction;
 import java.util.stream.Stream;
 
 public class OpTk {
@@ -297,5 +303,98 @@ public class OpTk {
 
     public static Stream<OpWrapper<?>> wrappedRootOpStream(MethodHandles.Lookup lookup, CoreOp.FuncOp op) {
         return RootSet.rootsWithoutVarFuncDeclarationsOrYields(lookup,op.bodies().getFirst().entryBlock());
+    }
+
+    public static FuncOpWrapper transformInvokes(MethodHandles.Lookup lookup, CoreOp.FuncOp funcOp, WrappedInvokeOpTransformer wrappedOpTransformer) {
+        return OpWrapper.wrap(lookup,funcOp.transform((b, op) -> {
+            if (op instanceof JavaOp.InvokeOp invokeOp) {
+                wrappedOpTransformer.apply(b, OpWrapper.wrap(lookup,invokeOp));
+            } else {
+                b.op(op);
+            }
+            return b;
+        }));
+    }
+
+    public interface WrappedInvokeOpTransformer extends BiFunction<Block.Builder, InvokeOpWrapper, Block.Builder> {
+        Block.Builder apply(Block.Builder block, InvokeOpWrapper op);
+    }
+
+    public static class ParamTable {
+        public static class Info {
+            public final int idx;
+            public final Block.Parameter parameter;
+            public final JavaType javaType;
+            public final CoreOp.VarOp varOp;
+            public GroupLayout layout = null;
+            public Class<?> clazz = null;
+            Info(int idx, Block.Parameter parameter, CoreOp.VarOp varOp) {
+                this.idx = idx;
+                this.parameter = parameter;
+                this.javaType = (JavaType) parameter.type();
+                this.varOp = varOp;
+            }
+            public boolean isPrimitive() {
+                return javaType instanceof PrimitiveType;
+            }
+            public void setLayout(GroupLayout layout) {
+                this.layout = layout;
+            }
+            public void setClass(Class<?> clazz) {
+                this.clazz = clazz;
+            }
+        }
+        final public BiMap<Block.Parameter, CoreOp.VarOp> parameterVarOpMap = new BiMap<>();
+        final public BiMap<Block.Parameter, JavaOp.InvokeOp> parameterInvokeOpMap = new BiMap<>();
+
+        final private Map<Block.Parameter, Info> parameterToInfo = new LinkedHashMap<>();
+        final private Map<CoreOp.VarOp, Info> varOpToInfo = new LinkedHashMap<>();
+
+        final private List<Info> list = new ArrayList<>();
+
+        public ParamTable.Info info(Block.Parameter parameter) {
+            return parameterToInfo.get(parameter);
+        }
+
+        void add(Map.Entry<Block.Parameter, CoreOp.VarOp> parameterToVarOp) {
+            //We add a new ParameterInfo to both maps using parameter and varOp as keys
+            varOpToInfo.put(parameterToVarOp.getValue(),
+                    // always called but convenient because computeIfAbsent returns what we added :)
+                    parameterToInfo.computeIfAbsent(parameterToVarOp.getKey(), (parameterKey) -> {
+                        var info = new ParamTable.Info(list.size(), parameterKey, parameterToVarOp.getValue());
+                        list.add(info);
+                        return info;
+                    })
+            );
+        }
+
+        public List<Info> list() {
+            return list;
+        }
+
+        public Stream<Info> stream() {
+            return list.stream();
+        }
+
+        final public  CoreOp funcOp;
+        ParamTable(CoreOp.FuncOp funcOp){
+            this.funcOp = funcOp;
+            funcOp.parameters().forEach(parameter -> {
+                Optional<Op.Result> optionalResult = parameter.uses().stream().findFirst();
+                optionalResult.ifPresentOrElse(result -> {
+                    var resultOp = result.op();
+                    if (resultOp instanceof CoreOp.VarOp varOp) {
+                        parameterVarOpMap.add(parameter, varOp);
+                        add(Map.entry(parameter, varOp));
+                    }else if (resultOp instanceof JavaOp.InvokeOp invokeOp) {
+                        parameterInvokeOpMap.add(parameter,invokeOp);
+                    }else{
+                        //System.out.println("neither varOp or an invokeOp "+resultOp.getClass().getName());
+                    }
+                }, () -> {
+                    throw new IllegalStateException("no use of param");
+                });
+            });
+        }
     }
 }
